% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalFunctions.R
\name{catDocument}
\alias{catDocument}
\title{Output with listed documents}
\usage{
catDocument(FullDocument, sep = "\\n", ...)
}
\arguments{
\item{FullDocument}{Document as structure by \code{\link{StructureDocument}}}

\item{sep}{The separation character(s) between each line.}

\item{...}{all extra arguments get passed along to the command "\code{\link[base]{cat}}"}
}
\description{
Behaves like \code{\link[base]{cat}}, but it first automatically unlists the exam to print the document.

Since the document is kept as a tree of lists, it simply abstract the idea of outputting the document. with one document.
}
\examples{
catDocument(TexExamRandomizer::testdoc)


}
