% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonparser.R
\name{compile_latex_directory}
\alias{compile_latex_directory}
\title{Compile LatexDirectory}
\usage{
compile_latex_directory(
  options = list(),
  input_directory,
  output_directory = input_directory,
  main_directory
)
}
\arguments{
\item{options}{See details}

\item{input_directory}{Input directory where the tex files are found}

\item{output_directory}{Output directory where  the pdf files will be placed}

\item{main_directory}{Directory we will compile from}
}
\value{
No return value
}
\description{
Internal function wrapping the options to compile a directory holding some text files into pdf files.
}
\details{
`options` is a list containing the possible options. The possible options are

\itemize{
\item compile: Should it compile the tex files
\item debug: Should we use debugging information. Otherwise, it will clean up the temporary files created by latex.
\item xelatex: Should it use the xelatex engine

}

The options can be defined directly given by using the other parameters to this function. (Note that those paramenters have precedence)
}
\keyword{internal}
