% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexmkCompilation.R
\name{CompileLatexDir}
\alias{CompileLatexDir}
\alias{CompileLatexDirEXAM}
\alias{CompileLatexDirHW}
\title{Compiling function}
\usage{
CompileLatexDir(pdf.dir.out, latex.dir.in, engine = "xelatex",
  compile.dir = NULL, extracmdoptions = NULL)

CompileLatexDirEXAM(pdf.dir.out, latex.dir.in, engine = "xelatex",
  compile.dir = NULL, extracmdoptions = NULL)

CompileLatexDirHW(pdf.dir.out, latex.dir.in, engine = "xelatex",
  compile.dir = NULL, extracmdoptions = NULL)
}
\arguments{
\item{pdf.dir.out}{Directory where the pdf output will be sent to}

\item{latex.dir.in}{Directory where all the tex files are found.}

\item{engine:}{Engine to use when compiling. Currently the options are \code{xelatex}, \code{lualatex}, \code{latex} and \code{pdflatex}

       \code{xelatex} is the default value. However, if the value is not recognized, \code{pdflatex} is used instead.}

\item{compile.dir:}{Directory from which compilation is invoked, if not specified, the directory we are compiling will be from where we do it. (This is specially usefull since we want to mantain the same relative paths from the main file).}
}
\value{
None
}
\description{
This function calls latexmk, which must be part of the system commands, a directory where tex files are found and outputs their pdf and other things in the pdf.dir.out
The functions \code{\link{CompileLatexDirEXAM}} and \code{\link{CompileLatexDirHW}} are identical wrappers of the same function, \code{\link{CompileLatexDir}}. Do not use them, they are just kept for "backwards" compatibility
}
\details{
Write the tex files relative paths to other files as to be read from the directory in which latex.dir.in is found
This function is intended to be use to compile a bunch of files which are stemmed from an original one. That is why the directory
}
\examples{

input_folder <- system.file(
    "extdata",
    "ExampleTexDocuments",
    package = "TexExamRandomizer")


TexExamRandomizer::CompileLatexDir(
    pdf.dir.out = tempdir(),
    engine= "pdf",
    latex.dir.in = input_folder,
    extracmdoptions = "-time")

}
\author{
Alejandro Recuenco \email{alejandrogonzalezrecuenco@gmail.com}
}
\keyword{internal}
