% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnswerSheets.R
\name{CountNumberOfSections}
\alias{CountNumberOfSections}
\title{CountNumberOfSections}
\usage{
CountNumberOfSections(Document)
}
\arguments{
\item{Document}{Document, as defined in \code{\link{StructureDocument}}. Remember however that the function \code{\link{StructureDocument}} returns the document and the preamble together in a list.}
}
\value{
A table, counting the number of "\code{\\cmdName}" items in which the document was divided when parsed for every begin-end section. It doesn't act resursively.

It will return a table with an integer that identifies the section, and a count, with how many items it found on that section. If it doesn't find any items or sections, it will return an empty table.
}
\description{
It counts the number of subparts in each section and outputs the result as a table. It doesn't act recursively, it only does the outermost layer.
}
\details{
The regular expression that defines this methods behaviour is the following

\code{"^[^[:digit:]]*([[:digit:]]+)_([^[:digit:]]+)_([[:digit:]]+)_(.*)"}

The replacement is simply \code{"\\1"}.

It tries to first find whether there exist attributes command and section that explain the command and section, before starting to use regexs on the names.
}
\seealso{
Other Extracting information: \code{\link{ConstructAnswerSheet}},
  \code{\link{FindExamAnswers}},
  \code{\link{GenerateShortAnswerSheet}}
}
\keyword{internal}
