% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuringdocument.R
\name{IsWellSectioned}
\alias{IsWellSectioned}
\title{IsWellSectioned}
\usage{
IsWellSectioned(u, v)
}
\arguments{
\item{u}{Vector, it assumes it is ordered in ascending ordered}

\item{v}{Vector, it assumes it is ordered in ascending ordered}
}
\value{
Logical value, TRUE if it is well ordered, FALSE it is not
}
\description{
Function to assure a set of sections is well sectioned.
}
\details{
Basically it makes sure that, \eqn{u[1]<v[1]<u[2]<v[2]}, etc
}
\seealso{
Other Structuring Document: \code{\link{CompileDocument}},
  \code{\link{DivideFile}}, \code{\link{FindStructure}},
  \code{\link{StructureDocument}}
}
\author{
Alejandro Recuenco \email{alejandrogonzalezrecuenco@gmail.com}
}
\keyword{internal}
