% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{testclass}
\alias{testclass}
\title{Sample class table}
\format{A dataframe with 5 rows and 4 columns}
\source{
self
}
\usage{
testclass
}
\description{
Sample class for testing with five students. The variables stored for each student are as follows
}
\details{
\itemize{
    \item Class
    \item Roll.Number
    \item Nickname
    \item Name
}
}
\keyword{datasets}
