% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{testdoc}
\alias{testdoc}
\title{Test document}
\format{A list with the format described in \code{\link{StructureDocument}}}
\source{
Created between me and my students in Suankularbwittayalai Rangsit School
}
\usage{
testdoc
}
\description{
A simple sample TeX document to test the package easily before deploying solutions.
}
\keyword{datasets}
