% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TexExamRandomizer.R
\docType{package}
\name{TexExamRandomizer-package}
\alias{TexExamRandomizer}
\alias{TexExamRandomizer-package}
\title{Generating Random Exams from 'LaTeX' documents}
\description{
This package is designed with exams and homework created in 'LaTeX' in mind. It allows to randomize and personalize exams and homework and it aids the user with grading them.
}
\details{
If you are using the exam class from 'LaTeX' already, it is likely that this program works as it is.

If you just want to randomize your exams,
\itemize{
    \item Look at \code{vignette("BasicUse", package = "TexExamRandomizer")} for an introduction of the concept behind this library and a quick way to start using it.

\item Look at \code{vignette("ExamOptions", package = "TexExamRandomizer")} for a more detailed explanations of what options can be used on a document.
}

If instead you are trying to use the library to create your own randomizer for a certain use you might have, you should start by looking at \code{\link{CreateRandomExams}} and \code{\link{GenerateHomework}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexrecuenco/TexExamRandomizer}
  \item Report bugs at \url{https://github.com/alexrecuenco/TexExamRandomizer/issues}
}

}
\author{
Alejandro Gonzalez Recuenco

e-mail: alejandrogonzalezrecuenco@gmail.com
}
